Initial commit for milestone 4
Updated readAllContacts to read from a comma delimited file into a BaseContact arrayList.
Updated saveAllContacts to save the current list as well as any updated via the console to the same file.
Created a new method in FAS that will split the array out to a string based on type
Added type to BaseContact to specify type of contact.